backend/requirements.txt
fastapi
uvicorn
python-multipart
moviepy
openai-whisper
torch
requests
